Move slf4j logger initialization to bootstrap module
And use a slf4j logger in the `ExtensionClassLoader` class
